requirements.txt
asgiref==3.6.0
certifi==2022.12.7
Django==4.1.5
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.5
django-nine==0.2.7
djangorestframework==3.14.0
elasticsearch==7.17.8
elasticsearch-dsl==7.4.0
packaging==23.0
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
